requirements.txt
asposestorage==1.0.2
basicsr==1.4.2
bigmodelvis==0.0.1
dashscope==1.24.0
dashscope==1.23.8
datasets==4.0.0
demjson3==3.0.6
docling==2.43.0
easydict==1.13
executor==23.2
fitz==0.0.1.dev2
Flask==3.1.2
hat==0.1.11
httpx==0.28.1
json5==0.12.1
llamaapi==0.1.36
lpips==0.1.4
manim==0.19.0
matplotlib==3.10.5
model==0.6.0
moviepy==1.0.3
natsort==8.4.0
numpy==2.3.2
openai==1.101.0
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
options==1.4.10
Pillow==11.3.0
protobuf==6.32.0
ptflops==0.7.5
pydub==0.25.1
pymol==3.2.0a0
python_docx==1.1.2
python_pptx==1.0.2
pytorch_msssim==1.0.0
PyYAML==6.0.2
PyYAML==6.0.2
qwen_vl_utils==0.0.11
regex==2024.11.6
reportlab==4.4.1
Requests==2.32.5
skimage==0.0
textblob==0.19.0
torch==2.9.0.dev20250801+cu128
torchvision==0.24.0.dev20250802+cu128
tqdm==4.67.1
transformers==4.55.0.dev0
typing_extensions==4.14.1
util==0.0.1
